+++ web-app/src/ng1/setup/setup.controller.js
@@ -1,7 +1,7 @@
 import { textField, linearProgress } from 'material-components-web';
 import zxcvbn from 'zxcvbn';
 
-var passwordStrengthMap = {
+const passwordStrengthMap = {
   0: {
     type: 'danger',
     text: 'Weak'
@@ -57,7 +57,7 @@ class SetupController {
   }
 
   onPasswordChange() {
-    var score = this.account.password && this.account.password.length ? zxcvbn(this.account.password, [this.account.username]).score : 0;
+    const score = this.account.password && this.account.password.length ? zxcvbn(this.account.password, [this.account.username]).score : 0;
     this.passwordStrengthScore = score + 1;
     this.passwordStrengthType = passwordStrengthMap[score].type;
     this.passwordStrength = passwordStrengthMap[score].text;
@@ -73,40 +73,35 @@ class SetupController {
       return;
     }
 
-    var index = this.pages.indexOf(this.page);
+    const index = this.pages.indexOf(this.page);
     this.page = this.pages[index + 1];
   }
 
   finish() {
     this._$http.post('/api/setup', this.account, {headers: { 'Content-Type': 'application/json' }}).success(() => {
-      // login the user
+      // Login the user after setup is complete
       this._UserService.signin({username: this.account.username, password: this.account.password}).then(response => {
-        var user = response.user;
-
-        this._UserService.authorize('local', user, false, {uid: this.account.uid}).success(data => {
-          if (data.device.registered) {
-            this.onSetupComplete({device: data});
-          }
-        });
+        this._UserService.authorize(response.token, this.account.uid).success(data => {
+          this.onSetupComplete({device: data});
+        })
       }, response => {
         this.showStatus = true;
         this.statusTitle = 'Error signing in';
         this.statusMessage = response.data || 'Please check your username and password and try again.';
         this.statusLevel = 'alert-danger';
       });
-    }).error(function() {
-    });
+    }).error(function() {});
   }
 }
 
 SetupController.$inject = ['$http', '$element', 'UserService'];
 
-var template = require('./setup.html');
-var bindings = {
+const template = require('./setup.html');
+const bindings = {
   api: '<',
   onSetupComplete: '&'
 };
-var controller = SetupController;
+const controller = SetupController;
 
 export {
   template,